Upgrade: 数据库模块使用 `async_scoped_session` 管理会话
- 数据库模块使用 `async_scoped_session` 管理会话
